second commit

+++ IRQ/Inc/CAN_IRQHandler.h
@@ -4,8 +4,9 @@
 #include "EXTI_IRQHandler.h"
 #include "main.h"
 #include "can.h"
+#include "fsm.h"
 
 /* CAN 报文打包发送接口 */
-void CAN_Send_Data(uint16_t id, uint8_t *data);
+void CAN_Send_Data(uint32_t id, uint8_t *data);
 
 #endif /* CAN_IRQHANDLER_H */
